+++ NEWS.md
@@ -1,3 +1,9 @@
+# dhcurve 0.2
+
+* Het bruikbaar interval vervalt bij afgeleide curves, deze worden nu
+gemodelleerd op basis van alle metingen binnen het interval 0.5-2.4 m
+(zie [issue 53](https://github.com/inbo/dhcurve/issues/53#issuecomment-1109585295))
+
 # dhcurve 0.1
 
 * Added a `NEWS.md` file to track changes to the package.